requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytz>=2024.1
pyjwt>=2.9.0
cryptography>=42.0.0
bcrypt>=4.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
anyio[trio]>=4.0.0
